requirements.txt
clickhouse-driver
PyYAML
pandas
schedule
requests